Squashed commit of the following:
commit e873dc653ecffc9cf86dab574eb0b54f0c1d15ab
Author: Antonio Jiménez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 19:10:16 2023 +0100

    Fixed imports

commit e36a7c9d91561028d94ce242f421898119091eb3
Author: Antonio Jiménez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 18:58:28 2023 +0100

    Fixed calls of GetEquationsForSolution

commit a7e259cd918d12bb44c683993e6309fca2492c3f
Author: Antonio Jimenez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 17:56:41 2023 +0100

    Update in GetEquationsForSolution
      - Added method "add_constants" to DRings
      - Added method "is_normal_form" to DPolynomial
      - Added method "change_ring" to DPolynomial

commit 5103f5af5a17168e35da80b52a8153181e872a6a
Author: Antonio Jimenez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 11:38:44 2023 +0100

    Refactored Grunbaum_Commutator for easier reading

commit 82898ef35887325f36778f3c9be428155322cdaf
Author: Antonio Jimenez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 11:06:11 2023 +0100

    Fixed linting errors

commit e256a1c91c70a5b16fc7a70beed32ead0ecec80e
Author: Antonio Jimenez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 10:59:36 2023 +0100

    Fixed tests and small typos

commit 236d8a25a01a667d5013fdbed616f6ea591cf5e8
Author: Antonio Jimenez-Pastor <kethewal@gmail.com>
Date:   Wed Nov 1 10:26:31 2023 +0100

    Refactor of module "hierarchies"
      - Changed name "hierarchies" -> "commutators"
      - Split "hierarchies" into 4 submodules:
        * almost_commuting: do the computation of Wilsons theorm
        * ideals: analyze algebraic ideals
        * commutators: compute commutators for specific types of operators
        * spectral: contains computations on spectral curves for commuting pairs of operators.
      - Refactored also documentation source code
      - Checked changes through notebooks and scripts
      - Updated version -> 0.0.4.5+is-34
