--- dalgebra/hierarchies/__init__.py
@@ -1,10 +1,0 @@
-r'''
-    Module with all the modules and algorithms to compute the integrable hierarchies.
-
-    TODO: 
-    
-    * add some explanation on what the integrable hierarchies are.
-    * add some referecence to the integrable hierarchies.
-'''
-
-from .hierarchies import *
