+++ docsrc/source/commutators/commutator.rst
@@ -1,0 +1,9 @@
+.. nodoctest
+
+Obtaining non-trivial centralizers for Specific Linear Differential Operators
+==============================================================================================================================
+
+.. automodule:: dalgebra.commutators.commutator
+   :members:
+   :undoc-members:
+   :show-inheritance:
