+++ docsrc/source/commutators/ideals.rst
@@ -1,0 +1,9 @@
+.. nodoctest
+
+Analyzing solutions to algebraic ideals
+==============================================================================================================================
+
+.. automodule:: dalgebra.commutators.ideals
+   :members:
+   :undoc-members:
+   :show-inheritance:
