+++ docsrc/source/commutators/spectral.rst
@@ -1,0 +1,9 @@
+.. nodoctest
+
+Spectral curves of commuting linear differential operators
+==============================================================================================================================
+
+.. automodule:: dalgebra.commutators.almost_commuting
+   :members:
+   :undoc-members:
+   :show-inheritance:
